+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 2023-04-17
+
+### Added
+
+- `RepositoryHttp` second TS generic usefull for `responseAdapter`
+- `useRepositoryHttp` second TS generic for `RepositoryHttp`
+
 ## [1.0.0] - 2023-04-12
 
 ### Added
@@ -50,6 +57,7 @@ All notable changes to this project will be documented in this file.
 - `UrlBuilder` a class to build URLs through a template;
 - `RepositoryHttp` an implementation of `Repository` interface to fetch data through `HttpClient`.
 
+[1.0.1]: https://github.com/volverjs/data/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/volverjs/data/compare/v0.0.4...v1.0.0
 [0.0.4]: https://github.com/volverjs/data/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/volverjs/data/compare/v0.0.2...v0.0.3